+++ geci_caller/__init__.py
@@ -1,4 +1,4 @@
 """A template Python module"""
 
-__version__ = "1.0.1"
+__version__ = "2.1.0"
 from .cli import *  # noqa