+++ geci_caller/cli.py
@@ -29,7 +29,7 @@ def write_posterior_results(
         ),
     }
     response = requests.post(url, files=files)
-    response.raise_for_status()
+    print_status_code(response)
     json_data = response.json()
     df = pd.DataFrame(json_data)
     df.to_csv(output_path, index=False)
@@ -49,8 +49,7 @@ def plot_cumulative_cpue_series_by_season(
     extension = output_path.split(".")[-1]
     response = requests.post(url, files=files, data={"format": extension})
     write_response_content(output_path, response)
-    response.raise_for_status()
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -67,7 +66,7 @@ def plot_cumulative_series_cpue_by_flight(
     extension = output_path.split(".")[-1]
     response = requests.post(url, files=files, data={"format": extension})
     write_response_content(output_path, response)
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -87,7 +86,7 @@ def write_bootstrap_progress_intervals(
     response = requests.post(url, files=files, data=data)
     with open(output_path, "w") as out_file:
         json.dump(response.json(), out_file, indent=2)
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -104,8 +103,7 @@ def filter_by_method(
     }
     data = {"method": method}
     response = requests.post(url, files=files, data=data)
-    print(response.status_code)
-    response.raise_for_status()
+    print_status_code(response)
     content = response.json()
     df = pd.DataFrame(content)
     df.to_csv(output_path, index=False)
@@ -127,7 +125,7 @@ def write_aerial_monitoring(
     response = requests.post(url, files=files, data=data)
     with open(output_path, "w") as out_file:
         json.dump(response.json(), out_file, indent=2)
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -145,7 +143,7 @@ def write_population_status_from_mixed_methods(
         "second_method_status": (second_method_status, second_method_file_like, "application/json"),
     }
     response = requests.post(url, files=files)
-    print(response.status_code)
+    print_status_code(response)
     with open(output_path, "w") as out_file:
         json.dump(response.json(), out_file, indent=2)
 
@@ -166,7 +164,7 @@ def plot_comparative_yearly_cpue(
     extension = output_path.split(".")[-1]
     response = requests.post(url, files=files, data={"format": extension})
     write_response_content(output_path, response)
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -186,7 +184,7 @@ def plot_comparative_catch_curves(
     extension = output_path.split(".")[-1]
     response = requests.post(url, files=files, data={"format": extension})
     write_response_content(output_path, response)
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -204,7 +202,7 @@ def write_population_status(
             files={"file": f},
             data={"bootstrapping_number": bootstrapping_number},
         )
-    response.raise_for_status()
+    print_status_code(response)
     with open(output_path, "w") as out_file:
         json.dump(response.json(), out_file, indent=2)
 
@@ -228,7 +226,7 @@ def write_instantaneous_and_cumulative_cpue(
         stream=True,
         timeout=600,
     )
-    response.raise_for_status()
+    print_status_code(response)
     json_data = response.json()
     df = pd.DataFrame(json_data)
     df.to_csv(output_path, index=False)
@@ -258,7 +256,7 @@ def write_csv_probability(
         stream=True,
         timeout=600,
     )
-    response.raise_for_status()
+    print_status_code(response)
     json_data = response.json()
     df = pd.DataFrame(json_data)
     df.to_csv(output_path, index=False)
@@ -275,7 +273,7 @@ def write_probability_progress_figure(
     files = {"file": (input_path, file_like, "text/csv")}
     response = requests.post(url, files=files)
     write_response_content(output_path, response)
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -295,7 +293,7 @@ def plot_custom_cpue_vs_cum_captures(
     extension = output_path.split(".")[-1]
     response = requests.post(url, files=files, data={"format": extension})
     write_response_content(output_path, response)
-    print(response.status_code)
+    print_status_code(response)
     return response
 
 
@@ -310,8 +308,28 @@ def plot_cpue_vs_cum_captures(
     files = {"file": (input_path, file_like, "text/csv")}
     response = requests.post(url, files=files, data={"format": extension})
     write_response_content(output_path, response)
-    print(response.status_code)
+    print_status_code(response)
     return response
+
+
+@cli.command()
+def render_instantaneous_and_cumulative_cpue_time_series(
+    input_path: str = typer.Option(help="Path of input data"),
+    output_path: str = typer.Option(help="Path of figure to write"),
+):
+    url = "http://islasgeci.org:100/plot_cumulative_cpue_time_series"
+    input_file_like = read_file_as_buffer(input_path)
+    files = {"file": (input_path, input_file_like, "text/csv")}
+    extension = output_path.split(".")[-1]
+    response = requests.post(url, files=files, data={"format": extension})
+    write_response_content(output_path, response)
+    print_status_code(response)
+    return response
+
+
+def print_status_code(response):
+    print(response.status_code)
+    response.raise_for_status()
 
 
 @cli.command()